requirements.txt
requests
keyring
qrcode
PyQt5
Pillow
